+++ tests/test_tnsim.py
@@ -302,6 +302,21 @@ class TestTN(unittest.TestCase):
         value2 = tn_mbqc.expectation_value(random_op2, [0, 1])
         np.testing.assert_almost_equal(value1, value2)
 
+    def test_ccx(self):
+        nqubits = 4
+        depth = 6
+        for _ in range(10):
+            circuit = rc.get_rand_circuit(nqubits, depth)
+            circuit.ccx(0, 1, 2)
+            pattern = circuit.transpile()
+            pattern.minimize_space()
+            state = circuit.simulate_statevector()
+            tn_mbqc = pattern.simulate_pattern(backend="tensornetwork")
+            random_op3 = random_op(3)
+            value1 = state.expectation_value(random_op3, [0, 1, 2])
+            value2 = tn_mbqc.expectation_value(random_op3, [0, 1, 2])
+            np.testing.assert_almost_equal(value1, value2)
+
     def test_with_graphtrans(self):
         nqubits = 4
         depth = 6